Create 10xbam_GEO_GSE222987.sh
A script to pull all of the 'Original File Format' files from GEO that are 10x BAM format. Then 10x BAM is then used to convert to the fastq format necessary for CellRanger alignments using bamtofastq. For example- the 10x BAM for 1 sample generates 12 fastq files, in the format necessary for CellRanger alignment. Therefore, we uploaded the 16 10x BAM files instead of the 192 fastq files.